Update LavaSmolderFade_WS2812B_FINAL.ino
In this modified version of the code, we've added a new variable called previousMillis that keeps track of the last time the LED strip was updated. In the loop, we use the millis() function to get the current time and check if it has been longer than BOIL_TIME since the last LED update. If it has, we generate the random boiling effect and update the LED strip. If not, we continue looping without doing anything.

This approach allows the program to continue running while waiting for the time delay to pass, so it remains responsive to user input.