Object-oriented smolder code
In this version of the code, I've created a new class called LavaSmolder that contains all the functionality for controlling the LED strip and generating the random boiling effect. The class has a constructor that takes in the parameters needed to set up the LED strip and initializes the member variables.

The class also has a begin method that initializes the LED strip and sets all the pixels to off. The update method generates the boiling effect and updates the LED strip with the new colors, using a non-blocking delay to prevent the program from becoming unresponsive.

We've also added a private method called generateBoilingEffect that generates the random boiling effect by setting the color of each segment to a random shade of red.

At the bottom of the code, we've created an instance of the LavaSmolder class called lavaSmolder, and then called its begin method in the setup function. In the loop function, we call the update method of the lavaSmolder instance to update the LED strip with the new colors.